#fixes
Removing of the functionnality of custom icons because the feature need a lot of configuaration as a apache server, etc... So, I will maybe create a automatized service later or a API linked with the PlexMediaServer... Also, the script will now wait than Discord load before running (because a error happend when you run it while Discord wasn't loaded). In the next release, a graphical interface will be added for blacklisted librairies.